Fix bug causing errors to be cleared when using autoProgress (#578)
* Use setState callback when setting errors to use the most up to date values

When setting a value for an input, we re run validation for any fields
that have errors. This was reading from the errors state directly which
can be out of date if called in the middle of a render. By using the
setErrors callback syntax we will get the most up to date values.

I've also removed a line which skipped validating the current field in
this setValue call. I'm not sure why this existed in the first place.

* Add changeset

* Add test